+++ test/game-test.js
@@ -6,47 +6,39 @@ describe("Game", function() {
   describe("setup", function() {
     it("defaults to a board size of 19", function() {
       var game = new Game();
-      expect(game.boardSize).to.be.null;
-      game.setup();
       expect(game.boardSize).to.equal(19);
     });
 
     it("does not allow unknown scoring types", function() {
-      var game = new Game();
-      expect(function() { game.setup({ boardSize: 19, scoring: "area" }); }).to.not.throw(Error);
-      expect(function() { game.setup({ boardSize: 19, scoring: "territory" }); }).to.not.throw(Error);
-      expect(function() { game.setup({ boardSize: 19, scoring: "equivalence" }); }).to.not.throw(Error);
+      expect(function() { new Game({ boardSize: 19, scoring: "area" }); }).to.not.throw(Error);
+      expect(function() { new Game({ boardSize: 19, scoring: "territory" }); }).to.not.throw(Error);
+      expect(function() { new Game({ boardSize: 19, scoring: "equivalence" }); }).to.not.throw(Error);
 
-      expect(function() { game.setup({ boardSize: 19, scoring: "Area" }); }).to.throw(Error,  "Unknown scoring type: Area");
-      expect(function() { game.setup({ boardSize: 19, scoring: "terRitory" }); }).to.throw(Error,  "Unknown scoring type: terRitory");
-      expect(function() { game.setup({ boardSize: 19, scoring: "gibberish" }); }).to.throw(Error,  "Unknown scoring type: gibberish");
+      expect(function() { new Game({ boardSize: 19, scoring: "Area" }); }).to.throw(Error,  "Unknown scoring type: Area");
+      expect(function() { new Game({ boardSize: 19, scoring: "terRitory" }); }).to.throw(Error,  "Unknown scoring type: terRitory");
+      expect(function() { new Game({ boardSize: 19, scoring: "gibberish" }); }).to.throw(Error,  "Unknown scoring type: gibberish");
     });
 
     it("does not allow unknown ko rules", function() {
-      var game = new Game();
+      expect(function() { new Game({ boardSize: 19, koRule: "simple" }); }).to.not.throw(Error);
 
-      expect(function() { game.setup({ boardSize: 19, koRule: "simple" }); }).to.not.throw(Error);
-
-      expect(function() { game.setup({ boardSize: 19, koRule: "Simple" }); }).to.throw(Error, "Unknown ko rule: Simple");
-      expect(function() { game.setup({ boardSize: 19, koRule: "SIMPLE" }); }).to.throw(Error, "Unknown ko rule: SIMPLE");
-      expect(function() { game.setup({ boardSize: 19, koRule: "superko" }); }).to.throw(Error, "Unknown ko rule: superko");
-      expect(function() { game.setup({ boardSize: 19, koRule: "positional" }); }).to.throw(Error, "Unknown ko rule: positional");
-      expect(function() { game.setup({ boardSize: 19, koRule: "gibberish" }); }).to.throw(Error, "Unknown ko rule: gibberish");
+      expect(function() { new Game({ boardSize: 19, koRule: "Simple" }); }).to.throw(Error, "Unknown ko rule: Simple");
+      expect(function() { new Game({ boardSize: 19, koRule: "SIMPLE" }); }).to.throw(Error, "Unknown ko rule: SIMPLE");
+      expect(function() { new Game({ boardSize: 19, koRule: "superko" }); }).to.throw(Error, "Unknown ko rule: superko");
+      expect(function() { new Game({ boardSize: 19, koRule: "positional" }); }).to.throw(Error, "Unknown ko rule: positional");
+      expect(function() { new Game({ boardSize: 19, koRule: "gibberish" }); }).to.throw(Error, "Unknown ko rule: gibberish");
     });
 
     it("does not allow any explicitly-given values which are null or undefined", function() {
-      var game = new Game();
-
-      expect(function() { game.setup({boardSize: undefined}) }).to.throw(Error, "Game option boardSize must not be set as null or undefined");
-      expect(function() { game.setup({boardSize: null}) }).to.throw(Error, "Game option boardSize must not be set as null or undefined");
-      expect(function() { game.setup({koRule: null}) }).to.throw(Error, "Game option koRule must not be set as null or undefined");
+      expect(function() { new Game({boardSize: undefined}) }).to.throw(Error, "Game option boardSize must not be set as null or undefined");
+      expect(function() { new Game({boardSize: null}) }).to.throw(Error, "Game option boardSize must not be set as null or undefined");
+      expect(function() { new Game({koRule: null}) }).to.throw(Error, "Game option koRule must not be set as null or undefined");
     });
   });
 
   describe("playAt", function() {
     it("is true or false depending on move success", function() {
       var game = new Game();
-      game.setup();
 
       expect(game.playAt(5, 10)).to.be.true;
       expect(game.playAt(5, 11)).to.be.true;
@@ -55,7 +47,6 @@ describe("Game", function() {
 
     it("starts with black and alternates", function() {
       var game = new Game();
-      game.setup();
 
       game.playAt(5, 5);
       expect(game.intersectionAt(5, 5).value).to.equal("black");
@@ -68,7 +59,6 @@ describe("Game", function() {
   describe("intersectionAt", function() {
     it("returns the intersection for the given point", function() {
       var game = new Game();
-      game.setup();
 
       expect(game.intersectionAt(4, 5).value).to.equal("empty");
       expect(game.intersectionAt(4, 5).y).to.equal(4);
@@ -82,8 +72,7 @@ describe("Game", function() {
     });
 
     it("errors when trying to retrieve values outside of the board", function() {
-      var game = new Game();
-      game.setup({ boardSize: 9 });
+      var game = new Game({ boardSize: 9 });
 
       expect(function() { game.intersectionAt(5, 14); }).to.throw(Error, "Intersection at (5, 14) would be outside the board");
       expect(function() { game.intersectionAt(4, -1); }).to.throw(Error, "Intersection position cannot be negative, but was given (4, -1)");
@@ -93,7 +82,6 @@ describe("Game", function() {
   describe("capturing", function() {
     it("removes black stones from the board", function() {
       var game = new Game();
-      game.setup();
 
       game.playAt(0, 0); // b
       game.playAt(0, 1); // w
@@ -107,7 +95,6 @@ describe("Game", function() {
 
     it("removes white stones from the board", function() {
       var game = new Game();
-      game.setup();
 
       game.pass(); // b
       game.playAt(0, 0); // w
@@ -122,7 +109,6 @@ describe("Game", function() {
 
     it("removes multi-stone groups from the board", function() {
       var game = new Game();
-      game.setup();
 
       game.playAt(0, 0); // b
       game.playAt(0, 1); // w
@@ -140,7 +126,6 @@ describe("Game", function() {
 
     it("removes multi-stone groups that share a liberty, counting unique spaces", function() {
       var game = new Game();
-      game.setup();
 
       game.playAt(0, 1); // b
       game.playAt(0, 2); // w
@@ -161,7 +146,6 @@ describe("Game", function() {
   describe("undo", function() {
     it("removes the last played stone", function() {
       var game = new Game();
-      game.setup();
 
       game.playAt(5, 10); // b
       game.playAt(5, 11); // w
@@ -182,7 +166,6 @@ describe("Game", function() {
   describe("isOver", function() {
     it("is false until two successive passes", function() {
       var game = new Game();
-      game.setup();
 
       expect(game.isOver()).to.be.false;
 
@@ -200,7 +183,6 @@ describe("Game", function() {
   describe("toggleDeadAt", function() {
     it("toggles stones dead as part of the scoring calculation", function() {
       var game = new Game();
-      game.setup();
 
       game.playAt(5, 5);
       game.playAt(5, 6);
@@ -223,7 +205,6 @@ describe("Game", function() {
   describe("handicap stones", function() {
     it("defaults to no handicap stones", function() {
       var game = new Game();
-      game.setup();
 
       expect(game.handicapStones).to.equal(0);
       expect(game.currentPlayer()).to.equal("black");
@@ -234,8 +215,7 @@ describe("Game", function() {
 
     it("allows 2-9 handicap stones", function() {
       [2, 3, 4, 5, 6, 7, 8, 9].forEach(function(h) {
-        var game = new Game();
-        game.setup({ handicapStones: h });
+        var game = new Game({ handicapStones: h });
 
         expect(game.handicapStones).to.equal(h);
         var nonEmptyPoints = game.currentState().intersections.filter(i => !i.isEmpty());
@@ -246,23 +226,22 @@ describe("Game", function() {
 
     it("does not allow invalid handicap stone values", function() {
       var game = new Game();
-      expect(function() { game.setup({ handicapStones: -1 }); }).to.throw(Error, "Only 2 to 9 handicap stones are supported");
-      expect(function() { game.setup({ handicapStones: 1 }); }).to.throw(Error, "Only 2 to 9 handicap stones are supported");
-      expect(function() { game.setup({ handicapStones: 10 }); }).to.throw(Error, "Only 2 to 9 handicap stones are supported");
+      expect(function() { new Game({ handicapStones: -1 }); }).to.throw(Error, "Only 2 to 9 handicap stones are supported");
+      expect(function() { new Game({ handicapStones: 1 }); }).to.throw(Error, "Only 2 to 9 handicap stones are supported");
+      expect(function() { new Game({ handicapStones: 10 }); }).to.throw(Error, "Only 2 to 9 handicap stones are supported");
     });
 
     it("does not allow handicap stones on non-standard sizes", function() {
       var game = new Game();
-      expect(function() { game.setup({ boardSize: 19, handicapStones: 2 }); }).to.not.throw(Error);
+      expect(function() { new Game({ boardSize: 19, handicapStones: 2 }); }).to.not.throw(Error);
 
       [3, 5, 7, 11, 15, 17, 21].forEach(b => {
-        expect(function() { game.setup({ boardSize: b, handicapStones: 2 }); }).to.throw(Error, "Handicap stones not supported on sizes other than 9x9, 13x13 and 19x19");
+        expect(function() { new Game({ boardSize: b, handicapStones: 2 }); }).to.throw(Error, "Handicap stones not supported on sizes other than 9x9, 13x13 and 19x19");
       });
     });
 
     it("treats handicap stone positions as illegal moves", function() {
-      var game = new Game();
-      game.setup({ handicapStones: 9 });
+      var game = new Game({ handicapStones: 9 });
 
       expect(game.isIllegalAt(3, 3)).to.be.true;
     });
@@ -271,7 +250,6 @@ describe("Game", function() {
   describe("komi", function() {
     it("does not add komi by default", function() {
       var game = new Game();
-      game.setup();
 
       game.pass();
       game.pass();
@@ -284,8 +262,7 @@ describe("Game", function() {
 
     it("supports integer and non-integer komi", function() {
       [0.5, 7, 7.5, 100].forEach(function(komiValue) {
-        var game = new Game();
-        game.setup({ komi: komiValue });
+        var game = new Game({ komi: komiValue });
 
         game.pass();
         game.pass();
@@ -299,7 +276,6 @@ describe("Game", function() {
   describe("coordinatesFor", function() {
     it("is the A19-T1 coordinate value for a point", function() {
       var game = new Game();
-      game.setup();
       expect(game.coordinatesFor(5, 3)).to.equal("D14");
       expect(game.coordinatesFor(0, 0)).to.equal("A19");
       expect(game.coordinatesFor(18, 18)).to.equal("T1");
@@ -307,7 +283,6 @@ describe("Game", function() {
 
     it("skips over I", function() {
       var game = new Game();
-      game.setup();
       expect(game.coordinatesFor(9, 7)).to.equal("H10");
       expect(game.coordinatesFor(9, 8)).to.equal("J10");
       expect(game.coordinatesFor(9, 9)).to.equal("K10");
@@ -317,7 +292,6 @@ describe("Game", function() {
   describe("move state", function() {
     it("is immutable", function() {
       var game = new Game();
-      game.setup();
       game.playAt(2, 3);
 
       expect(game.currentState().playedPoint.y).to.equal(2);
@@ -329,7 +303,6 @@ describe("Game", function() {
   describe("suicide restrictions", function() {
     it("prevents suicide", function() {
       var game = new Game();
-      game.setup();
 
       // in the corner
       expect(game.playAt(0, 1)).to.be.true; // b